+++ src/Application/ygo.application/Commands/UpdateMonsterCard/UpdateMonsterCardCommand.cs
@@ -6,7 +6,7 @@ namespace ygo.application.Commands.UpdateMonsterCard
     public class UpdateMonsterCardCommand : IRequest<CommandResult>
     {
         public long Id { get; set; }
-        public int? CardNumber { get; set; }
+        public long? CardNumber { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public int? CardLevel { get; set; }
