+++ src/Application/ygo.application/Commands/UpdateSpellCard/UpdateSpellCardCommand.cs
@@ -6,7 +6,7 @@ namespace ygo.application.Commands.UpdateSpellCard
     public class UpdateSpellCardCommand : IRequest<CommandResult>
     {
         public long Id { get; set; }
-        public int? CardNumber { get; set; }
+        public long? CardNumber { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public List<int> SubCategoryIds { get; set; }
